requirements.txt
streamlit
pandas
numpy
sklearn
matplotlib.pyplot
seaborn
statsmodels